Refactor categorie to category and remove servies
