+++ src/main/java/com/snj/snjback/Datafiller.java
@@ -1,26 +1,19 @@
 package com.snj.snjback;
 
-import com.snj.snjback.documents.Address;
-import com.snj.snjback.documents.Category;
-import com.snj.snjback.documents.Donation;
-import com.snj.snjback.documents.User;
-import com.snj.snjback.repositories.CategorieRepository;
+import com.snj.snjback.repositories.CategoryRepository;
 import com.snj.snjback.repositories.DonationRepository;
 import com.snj.snjback.repositories.UserRepository;
-import org.bson.types.ObjectId;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.stereotype.Component;
-
-import java.time.LocalDate;
 
 @Component
 public class Datafiller implements InitializingBean {
 
     private final UserRepository userRepository;
     private final DonationRepository donationRepository;
-    private final CategorieRepository categorieRepository;
+    private final CategoryRepository categorieRepository;
 
-    public Datafiller(UserRepository userRepository, DonationRepository donationRepository, CategorieRepository categorieRepository) {
+    public Datafiller(UserRepository userRepository, DonationRepository donationRepository, CategoryRepository categorieRepository) {
         this.userRepository = userRepository;
         this.donationRepository = donationRepository;
         this.categorieRepository = categorieRepository;
