+++ src/main/java/com/snj/snjback/services/CategoryService.java
@@ -1,4 +1,4 @@
-package com.snj.snjback.servies;
+package com.snj.snjback.services;
 
 import com.snj.snjback.documents.Category;
 import com.snj.snjback.exeption.ElementAlreadyExistsException;
@@ -6,18 +6,18 @@ import com.snj.snjback.exeption.ElementNotFoundException;
 import com.snj.snjback.forms.CategorieForm;
 import com.snj.snjback.mappers.CategorieMapper;
 import com.snj.snjback.documents.dto.CategorieDTO;
-import com.snj.snjback.repositories.CategorieRepository;
+import com.snj.snjback.repositories.CategoryRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-public class CategorieService implements GeneralService<Long,CategorieForm,CategorieDTO> {
+public class CategoryService implements GeneralService<Long,CategorieForm,CategorieDTO> {
     private final CategorieMapper mapper;
-    private final CategorieRepository repository;
+    private final CategoryRepository repository;
 
-    public CategorieService(CategorieMapper mapper, CategorieRepository repository) {
+    public CategoryService(CategorieMapper mapper, CategoryRepository repository) {
         this.mapper = mapper;
         this.repository = repository;
     }
